agentcore_strands/requirements.txt
bedrock-agentcore
strands-agents>=1.0.0
requests
boto3
